+++ tests/unit/test_folders.py
@@ -76,3 +76,83 @@ def test_create_folder_with_parent():
         with mock.patch("sasctl._services.folders.Folders.post"):
             with pytest.raises(ValueError):
                 folders.create_folder(FOLDER_NAME, parent="Doesnt Matter")
+
+
+@mock.patch("sasctl.core.Session.request")
+def test_get_folder_by_name(request):
+    """Verify that looking up a folder by name works."""
+    request.return_value.status_code = 200
+    request.return_value.json.return_value = {"name": "Spam"}
+
+    folder = folders.get_folder("Spam")
+    assert folder is not None
+    assert folder.name == "Spam"
+
+    # Service should do a search on folder name first followed by a get for folder details.
+    assert request.call_count == 2
+
+    url, params = request.call_args_list[0]
+    search_filter = params["params"]
+
+    # / should have been stripped out of folder name
+    assert '"Spam"' in search_filter
+
+
+@mock.patch("sasctl.core.Session.request")
+def test_get_folder_by_path(request):
+    """Verify that simple folder paths instead of names are handled."""
+    request.return_value.status_code = 200
+    request.return_value.json.return_value = {"name": "Spam"}
+
+    folder = folders.get_folder("/Spam")
+    assert folder is not None
+    assert folder.name == "Spam"
+
+    # Should not have called list_folders to search.  Should have been a direct call to lookup by path.
+    assert request.call_count == 1
+
+    url, params = request.call_args_list[0]
+    query_string = params["params"]
+
+    assert url[1] == "/folders/folders/@item"
+    assert query_string["path"] == "/Spam"
+
+
+@mock.patch("sasctl.core.Session.request")
+def test_get_folder_by_path(request):
+    """Verify that complex folder paths are handled."""
+    request.return_value.status_code = 200
+    request.return_value.json.return_value = {"name": "Spam"}
+
+    folder = folders.get_folder("Spam/Eggs/Spam")
+    assert folder is not None
+    assert folder.name == "Spam"
+
+    # Should not have called list_folders to search.  Should have been a direct call to lookup by path.
+    assert request.call_count == 1
+
+    url, params = request.call_args_list[0]
+    query_string = params["params"]
+
+    assert url[1] == "/folders/folders/@item"
+
+    # Leading "/" should have been added
+    assert query_string["path"] == "/Spam/Eggs/Spam"
+
+
+@mock.patch("sasctl.core.Session.request")
+def test_get_folder_by_delegate(request):
+    """Verify that folder can be found by using delegate strings (e.g. @public)."""
+    request.return_value.status_code = 200
+    request.return_value.json.return_value = {"name": "Spam"}
+
+    folder = folders.get_folder("@spam")
+    assert folder is not None
+    assert folder.name == "Spam"
+
+    # Should not have called list_folders to search.  Should have been a direct call to GET by id.
+    assert request.call_count == 1
+
+    url, params = request.call_args_list[0]
+    assert url[1] == "/folders/folders/@spam"
+    assert params == {}