+++ tests/unit/test_pageiterator.py
@@ -35,14 +35,17 @@ def paging(request):
             items=items[:start],
             count=len(items),
             links=[
-                {"rel": "next", "href": "/moaritems?start=%d&limit=%d" % (start, limit)}
+                {
+                    "rel": "next",
+                    "href": "/moaritems?start=%d&limit=%d" % (start, limit),
+                }
             ],
         )
 
         def side_effect(_, link, **kwargs):
             assert "limit=%d" % limit in link
             start = int(re.search(r"(?<=start=)[\d]+", link).group())
-            return RestObj(items=items[start : start + limit])
+            return RestObj(items=items[start: start + limit])
 
         req.side_effect = side_effect
         yield obj, items[:], req
@@ -53,43 +56,44 @@ def paging(request):
     assert req.call_count >= math.ceil(call_count)
 
 
-def test_no_paging_required():
-    """If "next" link not present, current items should be included."""
+@pytest.mark.incremental
+class TestPageIterator:
+    def test_no_paging_required(self):
+        """If "next" link not present, current items should be included."""
 
-    items = [{"name": "a"}, {"name": "b"}, {"name": "c"}]
-    obj = RestObj(items=items, count=len(items))
+        items = [{"name": "a"}, {"name": "b"}, {"name": "c"}]
+        obj = RestObj(items=items, count=len(items))
 
-    with mock.patch("sasctl.core.request") as request:
+        with mock.patch("sasctl.core.request") as request:
+            pager = PageIterator(obj)
+
+            # Returned page of items should preserve item order
+            items = next(pager)
+            for idx, item in enumerate(items):
+                assert item.name == RestObj(items[idx]).name
+
+        # No request should have been made to retrieve additional data.
+        try:
+            request.assert_not_called()
+        except AssertionError as e:
+            raise AssertionError(
+                f"method_calls={request.mock_calls}  call_args={request.call_args_list}"
+            )
+
+    def test_paging_required(self, paging):
+        """Requests should be made to retrieve additional pages."""
+        obj, items, _ = paging
+
         pager = PageIterator(obj)
+        init_count = pager._start
 
-        # Returned page of items should preserve item order
-        items = next(pager)
-        for idx, item in enumerate(items):
-            assert item.name == RestObj(items[idx]).name
-
-    # No request should have been made to retrieve additional data.
-    try:
-        request.assert_not_called()
-    except AssertionError as e:
-        raise AssertionError(
-            f"method_calls={request.mock_calls}  call_args={request.call_args_list}"
-        )
-
-
-def test_paging_required(paging):
-    """Requests should be made to retrieve additional pages."""
-    obj, items, _ = paging
-
-    pager = PageIterator(obj)
-    init_count = pager._start
-
-    for i, page in enumerate(pager):
-        for j, item in enumerate(page):
-            if i == 0:
-                item_idx = j
-            else:
-                # Account for initial page size not necessarily being same size
-                # as additional pages
-                item_idx = init_count + (i - 1) * pager._limit + j
-            target = RestObj(items[item_idx])
-            assert item.name == target.name
+        for i, page in enumerate(pager):
+            for j, item in enumerate(page):
+                if i == 0:
+                    item_idx = j
+                else:
+                    # Account for initial page size not necessarily being same size
+                    # as additional pages
+                    item_idx = init_count + (i - 1) * pager._limit + j
+                target = RestObj(items[item_idx])
+                assert item.name == target.name